requirements.txt
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysqlclient==1.4.5
pytz==2019.3
sqlparse==0.3.0
